kosm_dot: Fix purgeable_state type for atomic_get compatibility
Change purgeable_state from uint8 to int32 so atomic_get() works
without cast on both x86 and ARM64. Removes (int32*) casts in
kosm_dot_soft_fault and kosm_dot_purge_volatile.

+++ headers/private/kernel/vm/kosm_dot.h
@@ -197,7 +197,7 @@ struct KosmDot {
 	uint32				protection;		/* KOSM_PROT_* */
 	uint32				flags;			/* KOSM_DOT_* creation flags */
 	uint8				cache_policy;	/* KOSM_CACHE_* */
-	uint8				purgeable_state;/* KOSM_PURGE_* (atomic for reads) */
+	int32				purgeable_state;/* KOSM_PURGE_* (atomic for reads) */
 	uint32				tag;			/* KOSM_TAG_* */
 	uint32				wiring;			/* derived B_NO_LOCK / B_FULL_LOCK / B_CONTIGUOUS */
